# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CovidManagmentSystem",
    "version": "v1"
  },
  "paths": {
    "/patients": {
      "put": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/route": {
      "put": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientVisitedSite"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientVisitedSite"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientVisitedSite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientEncounterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientEncounterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientEncounterResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewPatientVisitedSite"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewPatientVisitedSite"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewPatientVisitedSite"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/encounter": {
      "put": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientEncounter"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientEncounter"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientEncounter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientEncounterResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientEncounterResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientEncounterResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/full": {
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PatientMedicalFile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientMedicalFile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientMedicalFile"
                }
              }
            }
          }
        }
      }
    },
    "/patients/new": {
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "parameters": [
          {
            "name": "dateTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/potential": {
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/isolated": {
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientEncounter"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/potential/{potentialPatientId}": {
      "post": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "parameters": [
          {
            "name": "potentialPatientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NewPatientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              }
            }
          }
        }
      }
    },
    "/labtests": {
      "post": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabTestResult"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LabTestResult"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LabTestResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewPatientResponse"
                }
              }
            }
          }
        }
      }
    },
    "/statistics": {
      "get": {
        "tags": [
          "CovidMgmtSystem"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CovidStatistics"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CovidStatistics"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CovidStatistics"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "houseNumber": {
            "type": "integer",
            "format": "int32"
          },
          "apartmentNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CityCovidStats": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "nullable": true
          },
          "infected": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CovidStatistics": {
        "type": "object",
        "properties": {
          "infected": {
            "type": "integer",
            "format": "int32"
          },
          "healed": {
            "type": "integer",
            "format": "int32"
          },
          "isolated": {
            "type": "integer",
            "format": "int32"
          },
          "cityStatistics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CityCovidStats"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LabTestResult": {
        "type": "object",
        "properties": {
          "labID": {
            "type": "string",
            "nullable": true
          },
          "testID": {
            "type": "string",
            "nullable": true
          },
          "patientID": {
            "type": "string",
            "nullable": true
          },
          "testDate": {
            "type": "string",
            "format": "date-time"
          },
          "isCovidPositive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "NewPatientEncounter": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewPatientEncounterResponse": {
        "type": "object",
        "additionalProperties": false
      },
      "NewPatientRequest": {
        "type": "object",
        "properties": {
          "govtID": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "houseResidentAmount": {
            "type": "integer",
            "format": "int32"
          },
          "isCovidPositive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "NewPatientResponse": {
        "type": "object",
        "properties": {
          "patientID": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewPatientVisitedSite": {
        "type": "object",
        "properties": {
          "dateOfVisit": {
            "type": "string",
            "format": "date-time"
          },
          "siteName": {
            "type": "string",
            "nullable": true
          },
          "siteAddress": {
            "$ref": "#/components/schemas/Address"
          }
        },
        "additionalProperties": false
      },
      "Patient": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "patientID": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "govID": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "houseResidentsAmount": {
            "type": "integer",
            "format": "int32"
          },
          "infectedByPatientID": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PatientEncounter": {
        "type": "object",
        "properties": {
          "potentialPatientDetails": {
            "$ref": "#/components/schemas/PotentialPatient"
          },
          "encounteredPatient": {
            "$ref": "#/components/schemas/Patient"
          }
        },
        "additionalProperties": false
      },
      "PatientMedicalFile": {
        "type": "object",
        "properties": {
          "patientDetails": {
            "$ref": "#/components/schemas/Patient"
          },
          "isCovidPositive": {
            "type": "boolean"
          },
          "labResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LabTestResult"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PotentialPatient": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "potentialPatientID": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}